+++ scalapb-runtime/src/main/scala/scalapb/options/MessageOptions.scala
@@ -33,6 +33,8 @@ package scalapb.options
   *   Adds a derives clause to the message case class
   * @param sealedOneofDerives
   *   Additional classes and traits to add to the derives clause of a sealed oneof.
+  * @param sealedOneofEmptyExtends
+  *   Additional traits to mixin for the empty case object of sealed oneofs.
   */
 @SerialVersionUID(0L)
 final case class MessageOptions(
@@ -48,6 +50,7 @@ final case class MessageOptions(
     sealedOneofCompanionExtends: _root_.scala.Seq[_root_.scala.Predef.String] = _root_.scala.Seq.empty,
     derives: _root_.scala.Seq[_root_.scala.Predef.String] = _root_.scala.Seq.empty,
     sealedOneofDerives: _root_.scala.Seq[_root_.scala.Predef.String] = _root_.scala.Seq.empty,
+    sealedOneofEmptyExtends: _root_.scala.Seq[_root_.scala.Predef.String] = _root_.scala.Seq.empty,
     unknownFields: _root_.scalapb.UnknownFieldSet = _root_.scalapb.UnknownFieldSet.empty
     ) extends scalapb.GeneratedMessage with scalapb.lenses.Updatable[MessageOptions] with _root_.scalapb.ExtendableMessage[MessageOptions] {
     @transient
@@ -101,6 +104,10 @@ final case class MessageOptions(
       sealedOneofDerives.foreach { __item =>
         val __value = __item
         __size += _root_.com.google.protobuf.CodedOutputStream.computeStringSize(12, __value)
+      }
+      sealedOneofEmptyExtends.foreach { __item =>
+        val __value = __item
+        __size += _root_.com.google.protobuf.CodedOutputStream.computeStringSize(13, __value)
       }
       __size += unknownFields.serializedSize
       __size
@@ -163,6 +170,10 @@ final case class MessageOptions(
         val __m = __v
         _output__.writeString(12, __m)
       };
+      sealedOneofEmptyExtends.foreach { __v =>
+        val __m = __v
+        _output__.writeString(13, __m)
+      };
       unknownFields.writeTo(_output__)
     }
     def clearExtends = copy(`extends` = _root_.scala.Seq.empty)
@@ -210,6 +221,10 @@ final case class MessageOptions(
     def addSealedOneofDerives(__vs: _root_.scala.Predef.String *): MessageOptions = addAllSealedOneofDerives(__vs)
     def addAllSealedOneofDerives(__vs: Iterable[_root_.scala.Predef.String]): MessageOptions = copy(sealedOneofDerives = sealedOneofDerives ++ __vs)
     def withSealedOneofDerives(__v: _root_.scala.Seq[_root_.scala.Predef.String]): MessageOptions = copy(sealedOneofDerives = __v)
+    def clearSealedOneofEmptyExtends = copy(sealedOneofEmptyExtends = _root_.scala.Seq.empty)
+    def addSealedOneofEmptyExtends(__vs: _root_.scala.Predef.String *): MessageOptions = addAllSealedOneofEmptyExtends(__vs)
+    def addAllSealedOneofEmptyExtends(__vs: Iterable[_root_.scala.Predef.String]): MessageOptions = copy(sealedOneofEmptyExtends = sealedOneofEmptyExtends ++ __vs)
+    def withSealedOneofEmptyExtends(__v: _root_.scala.Seq[_root_.scala.Predef.String]): MessageOptions = copy(sealedOneofEmptyExtends = __v)
     def withUnknownFields(__v: _root_.scalapb.UnknownFieldSet) = copy(unknownFields = __v)
     def discardUnknownFields = copy(unknownFields = _root_.scalapb.UnknownFieldSet.empty)
     def getFieldByNumber(__fieldNumber: _root_.scala.Int): _root_.scala.Any = {
@@ -226,6 +241,7 @@ final case class MessageOptions(
         case 10 => sealedOneofCompanionExtends
         case 11 => derives
         case 12 => sealedOneofDerives
+        case 13 => sealedOneofEmptyExtends
       }
     }
     def getField(__field: _root_.scalapb.descriptors.FieldDescriptor): _root_.scalapb.descriptors.PValue = {
@@ -243,6 +259,7 @@ final case class MessageOptions(
         case 10 => _root_.scalapb.descriptors.PRepeated(sealedOneofCompanionExtends.iterator.map(_root_.scalapb.descriptors.PString(_)).toVector)
         case 11 => _root_.scalapb.descriptors.PRepeated(derives.iterator.map(_root_.scalapb.descriptors.PString(_)).toVector)
         case 12 => _root_.scalapb.descriptors.PRepeated(sealedOneofDerives.iterator.map(_root_.scalapb.descriptors.PString(_)).toVector)
+        case 13 => _root_.scalapb.descriptors.PRepeated(sealedOneofEmptyExtends.iterator.map(_root_.scalapb.descriptors.PString(_)).toVector)
       }
     }
     def toProtoString: _root_.scala.Predef.String = _root_.scalapb.TextFormat.printToUnicodeString(this)
@@ -265,6 +282,7 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     val __sealedOneofCompanionExtends: _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String] = new _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String]
     val __derives: _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String] = new _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String]
     val __sealedOneofDerives: _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String] = new _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String]
+    val __sealedOneofEmptyExtends: _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String] = new _root_.scala.collection.immutable.VectorBuilder[_root_.scala.Predef.String]
     var `_unknownFields__`: _root_.scalapb.UnknownFieldSet.Builder = null
     var _done__ = false
     while (!_done__) {
@@ -295,6 +313,8 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
           __derives += _input__.readStringRequireUtf8()
         case 98 =>
           __sealedOneofDerives += _input__.readStringRequireUtf8()
+        case 106 =>
+          __sealedOneofEmptyExtends += _input__.readStringRequireUtf8()
         case tag =>
           if (_unknownFields__ == null) {
             _unknownFields__ = new _root_.scalapb.UnknownFieldSet.Builder()
@@ -315,6 +335,7 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
         sealedOneofCompanionExtends = __sealedOneofCompanionExtends.result(),
         derives = __derives.result(),
         sealedOneofDerives = __sealedOneofDerives.result(),
+        sealedOneofEmptyExtends = __sealedOneofEmptyExtends.result(),
         unknownFields = if (_unknownFields__ == null) _root_.scalapb.UnknownFieldSet.empty else _unknownFields__.result()
     )
   }
@@ -333,7 +354,8 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
         noDefaultValuesInConstructor = __fieldsMap.get(scalaDescriptor.findFieldByNumber(9).get).flatMap(_.as[_root_.scala.Option[_root_.scala.Boolean]]),
         sealedOneofCompanionExtends = __fieldsMap.get(scalaDescriptor.findFieldByNumber(10).get).map(_.as[_root_.scala.Seq[_root_.scala.Predef.String]]).getOrElse(_root_.scala.Seq.empty),
         derives = __fieldsMap.get(scalaDescriptor.findFieldByNumber(11).get).map(_.as[_root_.scala.Seq[_root_.scala.Predef.String]]).getOrElse(_root_.scala.Seq.empty),
-        sealedOneofDerives = __fieldsMap.get(scalaDescriptor.findFieldByNumber(12).get).map(_.as[_root_.scala.Seq[_root_.scala.Predef.String]]).getOrElse(_root_.scala.Seq.empty)
+        sealedOneofDerives = __fieldsMap.get(scalaDescriptor.findFieldByNumber(12).get).map(_.as[_root_.scala.Seq[_root_.scala.Predef.String]]).getOrElse(_root_.scala.Seq.empty),
+        sealedOneofEmptyExtends = __fieldsMap.get(scalaDescriptor.findFieldByNumber(13).get).map(_.as[_root_.scala.Seq[_root_.scala.Predef.String]]).getOrElse(_root_.scala.Seq.empty)
       )
     case _ => throw new RuntimeException("Expected PMessage")
   }
@@ -354,7 +376,8 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     noDefaultValuesInConstructor = _root_.scala.None,
     sealedOneofCompanionExtends = _root_.scala.Seq.empty,
     derives = _root_.scala.Seq.empty,
-    sealedOneofDerives = _root_.scala.Seq.empty
+    sealedOneofDerives = _root_.scala.Seq.empty,
+    sealedOneofEmptyExtends = _root_.scala.Seq.empty
   )
   implicit class MessageOptionsLens[UpperPB](_l: _root_.scalapb.lenses.Lens[UpperPB, scalapb.options.MessageOptions]) extends _root_.scalapb.lenses.ObjectLens[UpperPB, scalapb.options.MessageOptions](_l) {
     def `extends`: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Seq[_root_.scala.Predef.String]] = field(_.`extends`)((c_, f_) => c_.copy(`extends` = f_))
@@ -372,6 +395,7 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     def sealedOneofCompanionExtends: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Seq[_root_.scala.Predef.String]] = field(_.sealedOneofCompanionExtends)((c_, f_) => c_.copy(sealedOneofCompanionExtends = f_))
     def derives: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Seq[_root_.scala.Predef.String]] = field(_.derives)((c_, f_) => c_.copy(derives = f_))
     def sealedOneofDerives: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Seq[_root_.scala.Predef.String]] = field(_.sealedOneofDerives)((c_, f_) => c_.copy(sealedOneofDerives = f_))
+    def sealedOneofEmptyExtends: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Seq[_root_.scala.Predef.String]] = field(_.sealedOneofEmptyExtends)((c_, f_) => c_.copy(sealedOneofEmptyExtends = f_))
   }
   final val EXTENDS_FIELD_NUMBER = 1
   final val COMPANION_EXTENDS_FIELD_NUMBER = 2
@@ -385,6 +409,7 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
   final val SEALED_ONEOF_COMPANION_EXTENDS_FIELD_NUMBER = 10
   final val DERIVES_FIELD_NUMBER = 11
   final val SEALED_ONEOF_DERIVES_FIELD_NUMBER = 12
+  final val SEALED_ONEOF_EMPTY_EXTENDS_FIELD_NUMBER = 13
   def of(
     `extends`: _root_.scala.Seq[_root_.scala.Predef.String],
     companionExtends: _root_.scala.Seq[_root_.scala.Predef.String],
@@ -397,7 +422,8 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     noDefaultValuesInConstructor: _root_.scala.Option[_root_.scala.Boolean],
     sealedOneofCompanionExtends: _root_.scala.Seq[_root_.scala.Predef.String],
     derives: _root_.scala.Seq[_root_.scala.Predef.String],
-    sealedOneofDerives: _root_.scala.Seq[_root_.scala.Predef.String]
+    sealedOneofDerives: _root_.scala.Seq[_root_.scala.Predef.String],
+    sealedOneofEmptyExtends: _root_.scala.Seq[_root_.scala.Predef.String]
   ): _root_.scalapb.options.MessageOptions = _root_.scalapb.options.MessageOptions(
     `extends`,
     companionExtends,
@@ -410,7 +436,8 @@ object MessageOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     noDefaultValuesInConstructor,
     sealedOneofCompanionExtends,
     derives,
-    sealedOneofDerives
+    sealedOneofDerives,
+    sealedOneofEmptyExtends
   )
   // @@protoc_insertion_point(GeneratedMessageCompanion[scalapb.MessageOptions])
 }