+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [0.11.16]
 - sealed_oneof_companion_extends is handled also for optional sealed oneofs (#1639)
+- added sealed_oneof_empty_extends to add trait to the empty case object of
+  sealed oneofs (#1640)
 
 ## [0.11.15]
 - When more than one enum values share the same number, the first one would